player wont be able ot lose on first hit. process first hit func on game and a public re-arange func on minefield as first quick fix for this - too much shaerd responsability? I don't want the game to be responsible for re-arranging the board.
